+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [0.13.0] - 2024-05-07
 
 ### Added
 
@@ -104,6 +104,9 @@
 
 - Initial release
 
+[0.13.0]: https://github.com/tinted-theming/tinty/compare/v0.12.0...v0.13.0
+[0.12.0]: https://github.com/tinted-theming/tinty/compare/v0.11.0...v0.12.0
+[0.11.0]: https://github.com/tinted-theming/tinty/compare/v0.10.1...v0.11.0
 [0.10.1]: https://github.com/tinted-theming/tinty/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/tinted-theming/tinty/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/tinted-theming/tinty/compare/v0.8.1...v0.9.0